+++ CHANGES.rst
@@ -9,6 +9,9 @@ project endeavours to adhere to `Semantic Versioning`_.
 Unreleased
 ----------
 
+- ADDED: The ``authz``, ``identity``, and ``notification`` modules are now
+  exposed as public API on the ``annotator`` page global.
+
 2.0.0-alpha.2 - 2015-04-24
 --------------------------